notes controller user specific

+++ backend/src/controllers/notesController.js
@@ -1,7 +1,7 @@
 import Note from "../models/Note.js";
 import mongoose from "mongoose";
 
-export const getAllNotes = async (_, res) => {
+export const getAllNotes = async (req, res) => {
   try {
     const notes = await Note.find({ user: req.user._id })
       .sort({createdAt: -1});
@@ -42,7 +42,7 @@ export const createNote = async (req, res) => {
       content,
       user: req.user._id
     })
-    res.status(201).json({ message: "note created" }, newNote);
+    res.status(201).json(newNote);
   } catch (error) {
     console.log("error n createNote:", error);
     res.status(500).json({ message: "Server Error in CreateNote" });
@@ -52,14 +52,21 @@ export const createNote = async (req, res) => {
 export const updateNote = async (req, res) => {
   try {
     const { title, content } = req.body;
+
+     if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid Note ID" });
+    }
+
     const updatedNote = await Note.findOneAndUpdate(
       { _id: req.params.id, user: req.user._id },
       { title, content },
       { new: true }
     );
+
     if (!updatedNote)
       return res.status(404).json({ message: "Note not found" });
-    res.status(200).json({ message: "Note updated successfully" });
+
+    res.status(200).json(updatedNote);
   } catch (error) {
     console.log("error in updateNote:", error);
     res.status(500).json({ message: "Server Error" });
@@ -68,10 +75,12 @@ export const updateNote = async (req, res) => {
 
 export const deleteNote = async (req, res) => {
   try {
+     if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid Note ID" });
+    }
+    
     const deletedNote = await Note.findOneAndDelete(
-      {_id: req.params.id, user: req.user._id},
-      { title, content },
-      { new: true }
+      {_id: req.params.id, user: req.user._id}
     )
     if (!deletedNote)
       return res.status(404).json({ message: "Note not found" });
